utils and view clean up
