+++ src/infrastructure/aws_services.py
@@ -2,22 +2,7 @@
 import streamlit as st
 import datetime
 import boto3
-import jwt
-
-def assume_role_with_token(iam_token):
-    decoded_token = jwt.decode(iam_token, options={"verify_signature": False})
-    sts_client = boto3.client("sts", region_name="us-east-1")
-    response = sts_client.assume_role(
-        RoleArn=st.session_state.IAM_ROLE,
-        RoleSessionName="qapp",
-        ProvidedContexts=[
-            {
-                "ProviderArn": "arn:aws:iam::aws:contextProvider/IdentityCenter",
-                "ContextAssertion": decoded_token["sts:identity_context"],
-            }
-        ],
-    )
-    st.session_state.aws_credentials = response["Credentials"]
+from utils.sts_util import assume_role_with_token
     
 def get_qclient(idc_id_token: str):
     """
@@ -33,5 +18,5 @@ def get_qclient(idc_id_token: str):
         aws_secret_access_key=st.session_state.aws_credentials["SecretAccessKey"],
         aws_session_token=st.session_state.aws_credentials["SessionToken"],
     )
-    amazon_q = session.client("qbusiness", "us-east-1")
+    amazon_q = session.client("qbusiness", st.session_state.REGION)
     return amazon_q
