+++ src/utils/sts_util.py
@@ -2,14 +2,14 @@ import jwt
 import boto3
 import streamlit as st
 
-def assume_role_with_token(iam_token, iam_role, region):
+def assume_role_with_token(iam_token):
     """
     Assume IAM role using the IAM OIDC idToken.
     """
     decoded_token = jwt.decode(iam_token, options={"verify_signature": False})
-    sts_client = boto3.client("sts", region_name=region)
+    sts_client = boto3.client("sts", region_name=st.session_state.REGION)
     response = sts_client.assume_role(
-        RoleArn=iam_role,
+        RoleArn=st.session_state.IAM_ROLE,
         RoleSessionName="qapp",
         ProvidedContexts=[
             {
